saturation_point_

+++ rocksdb/plugin/zenfs/fs/zbd_zenfs.cc
@@ -2218,7 +2218,7 @@ void ZonedBlockDevice::WaitForOpenIOZoneToken(bool prioritized,WaitForOpenZoneCl
       }
       // else{
         priority_zone_resources_[open_class].wait(lk, [this,allocator_open_limit,open_class] {
-          uint64_t cur_open_classes=0;
+          int cur_open_classes=0;
           for(int oc = 0; oc<open_class; oc++){
             cur_open_classes+=cur_open_per_class_[open_class];
             if(cur_open_classes>saturation_point_){
